Moved radioactive obstacle.
