+++ Plugin Radioactive Cloud/src/main/java/net/simon987/pluginradioactivecloud/RadioactiveObstacle.java
@@ -1,0 +1,25 @@
+package net.simon987.pluginradioactivecloud;
+
+import net.simon987.server.GameServer;
+import net.simon987.server.game.objects.ControllableUnit;
+import net.simon987.server.game.objects.Enterable;
+import net.simon987.server.game.objects.GameObject;
+import net.simon987.server.game.objects.Radioactive;
+
+public class RadioactiveObstacle extends GameObject implements Radioactive, Enterable {
+
+	private final static int corruptionBlockSize = GameServer.INSTANCE.getConfig().getInt("radioactive_obstacle_corruption_block_size");
+	
+	@Override
+	public char getMapInfo() {
+		return 0;
+	}
+
+	@Override
+	public boolean enter(GameObject object) {
+		if (object instanceof ControllableUnit) {
+			((ControllableUnit) object).getCpu().getMemory().corrupt(corruptionBlockSize);
+		}
+		return false;
+	}
+}
